+++ src/client/locales/default.json
@@ -1681,5 +1681,6 @@
   "list_types": "List types",
   "Deposits": "Deposits",
   "Withdrawals": "Withdrawals",
+  "totals_calculated": "Totals can be calculated only for a defined from-till period.",
   "total": "TOTAL"
 }
