+++ src/client/locales/ru-RU.json
@@ -1681,5 +1681,6 @@
   "create_new_account_protect": "Чтобы защитить Hive от спама, свидетели блокчейна ввели небольшую единовременную плату за новые учетные записи Hive.",
   "Deposits": "Депозиты",
   "Withdrawals": "Вывод средств",
+  "totals_calculated": "Итоги могут быть рассчитаны только за определенный период от до.",
   "total": "ОБЩЕЕ"
 }
