+++ src/client/store/advancedReports/advancedReducer.js
@@ -6,6 +6,7 @@ const initialState = {
   isLoading: false,
   isLoadingMore: false,
   wallet: [],
+  loadingAllData: false,
 };
 
 export default function advancedReducer(state = initialState, action) {
@@ -14,6 +15,10 @@ export default function advancedReducer(state = initialState, action) {
       return {
         ...state,
         isLoading: true,
+        loadingAllData: true,
+        deposits: 0,
+        withdrawals: 0,
+        wallet: [],
       };
     }
     case GET_TRANSACTIONS_FOR_TABLE.SUCCESS: {
@@ -23,16 +28,16 @@ export default function advancedReducer(state = initialState, action) {
         ...state,
         wallet: data.wallet,
         hasMore: data.hasMore,
-        withdrawals: state.withdrawals + data.withdrawals,
-        deposits: state.deposits + data.deposits,
+        withdrawals: data.withdrawals,
+        deposits: data.deposits,
         isLoading: false,
         accounts: data.accounts,
+        loadingAllData: data.hasMore,
       };
     }
     case GET_MORE_TRANSACTIONS_FOR_TABLE.START: {
       return {
         ...state,
-        isLoadingMore: true,
       };
     }
     case GET_MORE_TRANSACTIONS_FOR_TABLE.SUCCESS: {
@@ -44,8 +49,8 @@ export default function advancedReducer(state = initialState, action) {
         hasMore: data.hasMore,
         withdrawals: state.withdrawals + data.withdrawals,
         deposits: state.deposits + data.deposits,
-        isLoadingMore: false,
         accounts: data.accounts,
+        loadingAllData: data.hasMore,
       };
     }
     default:
